made the visualizor render
